+++ src/readThroughPromiseCache.test.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2022-2023 Permanent Data Solutions, Inc. All Rights Reserved.
+ * Copyright (C) 2022-2024 Permanent Data Solutions, Inc. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -55,6 +55,25 @@ describe('ReadThroughPromiseCache Class', () => {
 
     expect(await cache.get('1')).to.equal('one');
     expect(await cache.get('1')).to.equal('one'); // the original cached value has not expired
+  });
+
+  it('should correctly show cache status with .getWithStatus method', async () => {
+    const testFunction = async (): Promise<string> => {
+      return 'always';
+    };
+    const cache = new ReadThroughPromiseCache<void, string>({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: testFunction,
+    });
+
+    expect(await cache.getWithStatus()).to.deep.equal({
+      status: 'miss',
+      data: 'always',
+    });
+    expect(await cache.getWithStatus()).to.deep.equal({
+      status: 'hit',
+      data: 'always',
+    }); // the original cached value has not expired
   });
 
   it('should throw error if readThroughFunction throws', async () => {
@@ -118,5 +137,86 @@ describe('ReadThroughPromiseCache Class', () => {
     expect(await cache.get('1')).to.equal('one');
     expect(await cache.get('2')).to.equal('two'); // over capacity
     expect(await cache.get('1')).to.equal('two'); // 1 is persisted in cache
+  });
+
+  it('should be able to manually put new entries in the cache', async () => {
+    const cache = new ReadThroughPromiseCache<string, string>({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: async () => 'one',
+    });
+
+    expect(await cache.get('1')).to.equal('one');
+    cache.put('1', Promise.resolve('two'));
+    expect(await cache.get('1')).to.equal('two');
+  });
+
+  it('should be able to remove entries from the cache', async () => {
+    const cache = new ReadThroughPromiseCache<string, string>({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: async () => 'one',
+    });
+
+    expect(await cache.get('1')).to.equal('one');
+    cache.put('1', Promise.resolve('two'));
+    expect(await cache.get('1')).to.equal('two');
+    cache.remove('1');
+    expect(await cache.get('1')).to.equal('one');
+  });
+
+  it('should be able to clear the cache', async () => {
+    const cache = new ReadThroughPromiseCache<string, string>({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: async () => 'one',
+    });
+
+    expect(cache.size()).to.equal(0);
+    expect(await cache.get('1')).to.equal('one');
+    expect(cache.size()).to.equal(1);
+    cache.clear();
+    expect(cache.size()).to.equal(0);
+  });
+
+  it('should be able to get the size of the cache', async () => {
+    const cache = new ReadThroughPromiseCache<string, string>({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: async () => 'one',
+    });
+
+    expect(cache.size()).to.equal(0);
+    expect(await cache.get('1')).to.equal('one');
+    expect(cache.size()).to.equal(1);
+  });
+
+  it('should pass readThroughData to readThroughFunction if provided', async () => {
+    let testTracker = 0;
+    const testFunction = async (
+      key: string,
+      readThroughData?: string,
+    ): Promise<string | undefined> => {
+      if (testTracker < 1) {
+        testTracker++;
+        return readThroughData;
+      } else {
+        return 'two';
+      }
+    };
+
+    const cache = new ReadThroughPromiseCache<
+      string,
+      string | undefined,
+      string | void
+    >({
+      cacheParams: { cacheCapacity: 10, cacheTTL: 60_000 },
+      readThroughFunction: testFunction,
+    });
+
+    expect(await cache.get('1', 'one')).to.equal('one');
+    expect(await cache.get('1', 'two')).to.equal('one');
+    cache.clear();
+    expect(await cache.get('1', 'three')).to.equal('two');
+    expect(await cache.get('1')).to.equal('two');
+    cache.clear();
+    testTracker = 0;
+    expect(await cache.get('1')).to.be.undefined;
   });
 });